+++ types/index.d.ts
@@ -21,29 +21,11 @@ export type AeroflyMissionPosition = {
 /**
  * State of aircraft systems. Configures power settings, flap positions etc
  */
-export type AeroflyMissionSetting =
-    | "cold_and_dark"
-    | "before_start"
-    | "taxi"
-    | "takeoff"
-    | "cruise"
-    | "approach"
-    | "landing"
-    | "winch_launch"
-    | "aerotow"
-    | "pushback";
+export type AeroflyMissionSetting = "cold_and_dark" | "before_start" | "taxi" | "takeoff" | "cruise" | "approach" | "landing" | "winch_launch" | "aerotow" | "pushback";
 /**
  * Types of checkpoints. Required are usually "origin", "departure_runway" at the start and "destination_runway", "destination" at the end.
  */
-export type AeroflyMissionCheckpointType =
-    | "origin"
-    | "departure_runway"
-    | "departure"
-    | "waypoint"
-    | "arrival"
-    | "approach"
-    | "destination_runway"
-    | "destination";
+export type AeroflyMissionCheckpointType = "origin" | "departure_runway" | "departure" | "waypoint" | "arrival" | "approach" | "destination_runway" | "destination";
 /**
  * Data for the aircraft to use on this mission
  * @property name lowercase Aerofly aircraft ID
@@ -203,46 +185,25 @@ export declare class AeroflyMission {
      * @param {AeroflyMissionConditions} [additionalAttributes.conditions] like time and weather for mission
      * @param {AeroflyMissionCheckpoint[]} [additionalAttributes.checkpoints] form the actual flight plan
      */
-    constructor(
-        title: string,
-        {
-            tutorialName,
-            description,
-            localizedTexts,
-            tags,
-            isFeatured,
-            difficulty,
-            flightSetting,
-            aircraft,
-            callsign,
-            origin,
-            destination,
-            distance,
-            duration,
-            isScheduled,
-            finish,
-            conditions,
-            checkpoints,
-        }?: {
-            tutorialName?: string | null;
-            description?: string;
-            localizedTexts?: AeroflyLocalizedText[];
-            tags?: string[];
-            isFeatured?: boolean | null;
-            difficulty?: number | null;
-            flightSetting?: AeroflyMissionSetting;
-            aircraft?: AeroflyMissionAircraft;
-            callsign?: string;
-            origin?: AeroflyMissionPosition;
-            destination?: AeroflyMissionPosition;
-            distance?: number | null;
-            duration?: number | null;
-            isScheduled?: boolean | null;
-            finish?: AeroflyMissionTargetPlane | null;
-            conditions?: AeroflyMissionConditions;
-            checkpoints?: AeroflyMissionCheckpoint[];
-        },
-    );
+    constructor(title: string, { tutorialName, description, localizedTexts, tags, isFeatured, difficulty, flightSetting, aircraft, callsign, origin, destination, distance, duration, isScheduled, finish, conditions, checkpoints, }?: {
+        tutorialName?: string | null;
+        description?: string;
+        localizedTexts?: AeroflyLocalizedText[];
+        tags?: string[];
+        isFeatured?: boolean | null;
+        difficulty?: number | null;
+        flightSetting?: AeroflyMissionSetting;
+        aircraft?: AeroflyMissionAircraft;
+        callsign?: string;
+        origin?: AeroflyMissionPosition;
+        destination?: AeroflyMissionPosition;
+        distance?: number | null;
+        duration?: number | null;
+        isScheduled?: boolean | null;
+        finish?: AeroflyMissionTargetPlane | null;
+        conditions?: AeroflyMissionConditions;
+        checkpoints?: AeroflyMissionCheckpoint[];
+    });
     /**
      * @returns {string} indexed checkpoints
      */
@@ -303,16 +264,7 @@ export declare class AeroflyMissionConditions {
      * @param {?number} [additionalAttributes.temperature] in °C, will overwrite thermalStrength
      * @param {AeroflyMissionConditionsCloud[]} [additionalAttributes.clouds] for the whole flight
      */
-    constructor({
-        time,
-        wind,
-        turbulenceStrength,
-        thermalStrength,
-        visibility,
-        visibility_sm,
-        temperature,
-        clouds,
-    }?: {
+    constructor({ time, wind, turbulenceStrength, thermalStrength, visibility, visibility_sm, temperature, clouds, }?: {
         time?: Date;
         wind?: {
             direction: number;
@@ -442,6 +394,12 @@ export declare class AeroflyMissionCheckpoint {
      */
     altitude: number;
     /**
+     * @property {?boolean} altitudeConstraint The altitude given in `altitude`
+     *    will be interpreted as mandatory flight plan altitude instead of
+     *    suggestion.
+     */
+    altitudeConstraint: boolean | null;
+    /**
      * @property {?number} direction of runway, in degree
      */
     direction: number | null;
@@ -476,38 +434,27 @@ export declare class AeroflyMissionCheckpoint {
      *    84 reference ellipsoid
      * @param {?number} [additionalAttributes.altitude_feet] The height in feet above or below the WGS
      *    84 reference ellipsoid. Will overwrite altitude
-     * @param {number} [additionalAttributes.direction] of runway, in degree
+     * @param {number} [additionalAttributes.altitudeConstraint] The altitude given in `altitude`
+     *    will be interpreted as mandatory flight plan altitude instead of
+     *    suggestion.
+     * @param {boolean} [additionalAttributes.direction] of runway, in degree
      * @param {?number} [additionalAttributes.slope] of runway
      * @param {?number} [additionalAttributes.length] of runway, in meters
      * @param {?number} [additionalAttributes.length_feet] of runway, in feet. Will overwrite length
      * @param {?number} [additionalAttributes.frequency] of runways or navigational aids, in Hz; multiply by 1000 for kHz, 1_000_000 for MHz
      * @param {?boolean} [additionalAttributes.flyOver] if waypoint is meant to be flown over
      */
-    constructor(
-        name: string,
-        type: AeroflyMissionCheckpointType,
-        longitude: number,
-        latitude: number,
-        {
-            altitude,
-            altitude_feet,
-            direction,
-            slope,
-            length,
-            length_feet,
-            frequency,
-            flyOver,
-        }?: {
-            altitude?: number;
-            altitude_feet?: number | null;
-            direction?: number | null;
-            slope?: number | null;
-            length?: number | null;
-            length_feet?: number | null;
-            frequency?: number | null;
-            flyOver?: boolean | null;
-        },
-    );
+    constructor(name: string, type: AeroflyMissionCheckpointType, longitude: number, latitude: number, { altitude, altitude_feet, altitudeConstraint, direction, slope, length, length_feet, frequency, flyOver, }?: {
+        altitude?: number;
+        altitude_feet?: number | null;
+        altitudeConstraint?: boolean | null;
+        direction?: number | null;
+        slope?: number | null;
+        length?: number | null;
+        length_feet?: number | null;
+        frequency?: number | null;
+        flyOver?: boolean | null;
+    });
     /**
      * @param {number} altitude_feet
      */